Split Knob into controlled and uncontrolled components
